distributes pipeline.py onto new dimcat.steps package and adds ._pipeline field to Dataset
