+++ src/dimcat/analyzers/base.py
@@ -8,10 +8,10 @@ from typing import Any, ClassVar, Iterable, Optional, Type, TypeVar
 import marshmallow as mm
 from dimcat.base import ObjectEnum
 from dimcat.dataset.processed import AnalyzedDataset
-from dimcat.pipeline import PipelineStep
 from dimcat.resources.base import DimcatResource, SomeSeries
 from dimcat.resources.features import Feature, FeatureSpecs
 from dimcat.resources.results import Result
+from dimcat.steps import PipelineStep
 
 logger = logging.getLogger(__name__)
