+++ src/dimcat/resources/__init__.py
@@ -1,6 +1,12 @@
 import logging
 
 from .base import DimcatIndex, DimcatResource, PieceIndex
-from .features import FeatureName, Notes
+from .features import (
+    Feature,
+    FeatureName,
+    FeatureSpecs,
+    Notes,
+    features_argument2config_list,
+)
 
 logger = logging.getLogger(__name__)
