+++ src/dimcat/steps/__init__.py
@@ -1,0 +1,7 @@
+import logging
+
+from .base import PipelineStep
+from .extractors import FeatureExtractor
+from .pipelines import Pipeline
+
+logger = logging.getLogger(__name__)
